federated_screener/federated_screener/requirements.txt
flwr[simulation]
tensorflow
web3
fastapi
uvicorn
python-dotenv
pandas
scikit-learn
eth-account
PyPDF2
